Added AddEdit Payment view and functionality Replaced WpfScheduler.dll with project to improve control flexibility Improved WpfScheduler by adding MouseEnter and MouseLeave events Added Events highlighting Improved WpfScheduler Event model to work with DSManager models Added LockedDates model with mapping Reverted LiveCharts framework version to 0.8.0 to make line chart work properly Changed Car Exploitation column chart to line chart Fixed some minor problems with statistics From now on statistics filtering is triggered automatically when filter parameter is changed From now on PDF files are saved to a different location to make them easily accessible From now on it's impossible to delete Boss account

+++ DSManager/ViewModel/Pages/UsersViewModel.cs
@@ -7,6 +7,7 @@ using GalaSoft.MvvmLight.Messaging;
 
 using DSManager.Messengers;
 using DSManager.Model.Entities;
+using DSManager.Model.Enums;
 using DSManager.Model.Services;
 using DSManager.View.Windows;
 
@@ -113,10 +114,14 @@ namespace DSManager.ViewModel.Pages {
                     if(_user == null) {
                         ShowDialog("Błąd", "Nie wybrano żadnego użytkownika!");
                     } else {
-                        if(await ConfirmationDialog("Potwierdź", "Czy jesteś pewien, że chcesz usunąć danego użytkownika?"))
-                            using(var repository = new BaseRepository()) {
-                                repository.Delete(_user);
-                            }
+                        if (_user.AccountType == AccountType.Boss) {
+                            ShowDialog("Błąd", "Nie można usunąć użytkownika z najwyższymi uprawnieniami!");
+                        } else {
+                            if(await ConfirmationDialog("Potwierdź", "Czy jesteś pewien, że chcesz usunąć danego użytkownika?"))
+                                using(var repository = new BaseRepository()) {
+                                    repository.Delete(_user);
+                                }
+                        }
                     }
                     FillUsers(_filter);
                 }));
